Improve module split in the kem combiner
