+++ mls-rs-crypto-hpke/src/lib.rs
@@ -10,8 +10,6 @@ wasm_bindgen_test::wasm_bindgen_test_configure!(run_in_browser);
 
 pub mod context;
 pub mod dhkem;
-#[cfg(not(mls_build_async))]
-pub mod ghp_kem_combiner;
 pub mod hpke;
 pub mod kdf;
 pub mod kem_combiner;
